Update statue-cave.md
Wind of Kron does actually blow in Statue Cave, with the first gust happening after 1,700 turns have elapsed. It's a pretty insignificant detail, but it can be proven in Puzzle 127: Peach stealing beam by placing a Fort. Staff wall in the way of the Cave Mamels and then consuming the Juicy Peach when your Food Gauge reaches 0, and then waiting the rest of the time.